requirements.txt
appnope==0.1.0
beautifulsoup4==4.6.0
bleach==2.1.2
bs4==0.0.1
certifi==2018.1.18
chardet==3.0.4
cltk==0.1.82
cycler==0.10.0
decorator==4.2.1
entrypoints==0.2.3
gitdb2==2.0.3
GitPython==2.1.8
html5lib==1.0.1
idna==2.6
ipykernel==4.8.0
ipython==6.2.1
ipython-genutils==0.2.0
ipywidgets==7.1.1
jedi==0.11.1
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.2
jupyter-console==5.2.0
jupyter-core==4.4.0
MarkupSafe==1.0
matplotlib==2.1.2
mistune==0.8.3
nbconvert==5.3.1
nbformat==4.4.0
nltk==3.2.5
notebook==5.4.0
numpy==1.14.0
pandas==0.22.0
pandocfilters==1.4.2
parso==0.1.1
pexpect==4.3.1
pickleshare==0.7.4
prompt-toolkit==1.0.15
ptyprocess==0.5.2
pyenchant==2.0.0
Pygments==2.2.0
pyparsing==2.2.0
python-crfsuite==0.9.5
python-dateutil==2.6.1
pytz==2017.3
pyuca==1.2
PyYAML==3.12
pyzmq==16.0.4
qtconsole==4.3.1
regex==2018.2.8
requests==2.18.4
scikit-learn==0.19.1
scipy==1.0.0
Send2Trash==1.4.2
simplegeneric==0.8.1
six==1.11.0
smmap2==2.0.3
terminado==0.8.1
testpath==0.3.1
tornado==4.5.3
tqdm==4.19.5
traitlets==4.3.2
Unidecode==1.0.22
urllib3==1.22
wcwidth==0.1.7
webencodings==0.5.1
Whoosh==2.7.4
widgetsnbextension==3.1.3
imblearn==0.3.3
